refactor: ui state 패키지 이동
- uiState  domain 패키지 -> ui 패키지로 이동
