+++ app/src/main/java/com/woowa/banchan/ui/common/uistate/UiState.kt
@@ -1,4 +1,4 @@
-package com.woowa.banchan.domain
+package com.woowa.banchan.ui.common.uistate
 
 sealed class UiState {
